backend/pyproject.toml
[project]
name = "aivideo-retriever-backend"
version = "1.0.0"
description = "AI Video Retriever Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic-settings==2.1.0",
    "python-multipart==0.0.6",
    "qdrant-client==1.7.0",
    "numpy>=1.26.0",
    "tqdm==4.66.1",
    "faiss-cpu>=1.7.4",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

